requirements.txt
DAWG-Python==0.7.1
Django==1.6.6
PyYAML==3.11
South==1.0
argparse==1.2.1
django-tastypie==0.11.1
docopt==0.6.2
mixer==4.8.1
pymorphy2==0.8
pymorphy2-dicts==2.4.390857.3662349
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
wsgiref==0.1.2
